+++ src/main/java/com/internproject/LatenSync/service/impl/UserServiceImpl.java
@@ -5,6 +5,8 @@ import com.internproject.LatenSync.exception.ResourceNotFoundException;
 import com.internproject.LatenSync.repository.UserRepository;
 import com.internproject.LatenSync.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -24,7 +26,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User addUser(User user) {
+        user.setUserName(user.getUserName());
         user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+        user.setUser_type(user.getUser_type());
         return userRepository.save(user);
     }
 
@@ -36,12 +40,31 @@ public class UserServiceImpl implements UserService {
     @Override
     public void removeUser(String userName) {
         User user = userRepository.findByUserName(userName);
-        userRepository.findById(user.getId()).orElseThrow(()-> new ResourceNotFoundException("Metrics","Id",userName));
+        if (user == null) {
+            throw new ResourceNotFoundException("User", "Username", userName);
+        }
         userRepository.deleteById(user.getId());
     }
+
 
     @Override
     public User getUserByUserName(String user_name) {
         return userRepository.findByUserName(user_name);
+    }
+
+    @Override
+    public boolean authenticate(String username, String password) {
+        User user = userRepository.findByUserName(username);
+
+        if (user == null) {
+            throw new UsernameNotFoundException("User not found in the database");
+        }
+
+        // BCrypt's matches method compares plaintext password with hashed password
+        if(!bCryptPasswordEncoder.matches(password, user.getPassword())) {
+            throw new BadCredentialsException("The password is incorrect");
+        }
+
+        return true;
     }
 }
